add calculateAverage and mapLengthOfWords
the calculateAverage function takes input in the form of array and
returns the average of the values in the array.
the mapLengthOfWords function takes input in the form of array. the
input array contains number of words and the function returns the number of characters of
every word in the form of array.